+++ display/download.go
@@ -1,6 +1,8 @@
 package display
 
 import (
+	"fmt"
+	"io"
 	"io/ioutil"
 	"net/url"
 	"os"
@@ -8,10 +10,216 @@ import (
 	"path/filepath"
 	"strconv"
 	"strings"
+	"time"
 
+	"github.com/gdamore/tcell"
 	"github.com/makeworld-the-better-one/amfora/config"
 	"github.com/makeworld-the-better-one/amfora/structs"
+	"github.com/makeworld-the-better-one/go-gemini"
+	"github.com/makeworld-the-better-one/progressbar/v3"
+	"github.com/spf13/viper"
+	"gitlab.com/tslocum/cview"
 )
+
+// For choosing between download and the portal - copy of YesNo basically
+var dlChoiceModal = cview.NewModal().
+	SetTextColor(tcell.ColorWhite).
+	SetText("That file could not be displayed. What would you like to do?").
+	AddButtons([]string{"Download", "Open in portal", "Cancel"})
+
+// Channel to indicate what choice they made using the button text
+var dlChoiceCh = make(chan string)
+
+var dlModal = cview.NewModal().
+	SetTextColor(tcell.ColorWhite)
+
+func dlInit() {
+	if viper.GetBool("a-general.color") {
+		dlChoiceModal.SetButtonBackgroundColor(tcell.ColorNavy).
+			SetButtonTextColor(tcell.ColorWhite).
+			SetBackgroundColor(tcell.ColorPurple)
+		dlModal.SetButtonBackgroundColor(tcell.ColorNavy).
+			SetButtonTextColor(tcell.ColorWhite).
+			SetBackgroundColor(tcell.Color130) // DarkOrange3, #af5f00
+	} else {
+		dlChoiceModal.SetButtonBackgroundColor(tcell.ColorWhite).
+			SetButtonTextColor(tcell.ColorBlack).
+			SetBackgroundColor(tcell.ColorBlack)
+		dlModal.SetButtonBackgroundColor(tcell.ColorWhite).
+			SetButtonTextColor(tcell.ColorBlack).
+			SetBackgroundColor(tcell.ColorBlack)
+	}
+
+	dlChoiceModal.SetBorder(true)
+	dlChoiceModal.SetBorderColor(tcell.ColorWhite)
+	dlChoiceModal.SetDoneFunc(func(buttonIndex int, buttonLabel string) {
+		dlChoiceCh <- buttonLabel
+	})
+	dlChoiceModal.GetFrame().SetTitleColor(tcell.ColorWhite)
+	dlChoiceModal.GetFrame().SetTitleAlign(cview.AlignCenter)
+
+	dlModal.SetBorder(true)
+	dlModal.SetBorderColor(tcell.ColorWhite)
+	dlModal.SetDoneFunc(func(buttonIndex int, buttonLabel string) {
+		if buttonLabel == "Ok" {
+			tabPages.SwitchToPage(strconv.Itoa(curTab))
+		}
+	})
+	dlModal.GetFrame().SetTitleColor(tcell.ColorWhite)
+	dlModal.GetFrame().SetTitleAlign(cview.AlignCenter)
+	dlModal.GetFrame().SetTitle(" Download ")
+}
+
+// dlChoice displays the download choice modal and acts on the user's choice.
+// It should run in a goroutine.
+func dlChoice(u string, resp *gemini.Response) {
+	defer resp.Body.Close()
+
+	parsed, err := url.Parse(u)
+	if err != nil {
+		Error("URL Error", err.Error())
+		return
+	}
+
+	tabPages.ShowPage("dlChoice")
+	tabPages.SendToFront("dlChoice")
+	App.SetFocus(dlChoiceModal)
+	App.Draw()
+
+	choice := <-dlChoiceCh
+	if choice == "Download" {
+		tabPages.HidePage("dlChoice")
+		App.Draw()
+		downloadURL(u, resp)
+		return
+	}
+	if choice == "Open in portal" {
+		// Open in mozz's proxy
+		portalURL := u
+		if parsed.RawQuery != "" {
+			// Remove query and add encoded version on the end
+			query := parsed.RawQuery
+			parsed.RawQuery = ""
+			portalURL = parsed.String() + "%3F" + query
+		}
+		portalURL = strings.TrimPrefix(portalURL, "gemini://") + "?raw=1"
+		handleHTTP("https://portal.mozz.us/gemini/"+portalURL, false)
+		tabPages.SwitchToPage(strconv.Itoa(curTab))
+		App.Draw()
+		return
+	}
+	tabPages.SwitchToPage(strconv.Itoa(curTab))
+	App.Draw()
+}
+
+// downloadURL pulls up a modal to show download progress and saves the URL content.
+// downloadPage should be used for Page content.
+func downloadURL(u string, resp *gemini.Response) {
+	_, _, width, _ := dlModal.GetInnerRect()
+	// Copy of progressbar.DefaultBytesSilent with custom width
+	bar := progressbar.NewOptions64(
+		-1,
+		progressbar.OptionSetWidth(width),
+		progressbar.OptionSetWriter(ioutil.Discard),
+		progressbar.OptionShowBytes(true),
+		progressbar.OptionThrottle(65*time.Millisecond),
+		progressbar.OptionShowCount(),
+		progressbar.OptionSpinnerType(14),
+	)
+	bar.RenderBlank()
+
+	savePath, err := downloadNameFromURL(u, "")
+	if err != nil {
+		Error("Download Error", "Error deciding on file name: "+err.Error())
+		return
+	}
+	f, err := os.OpenFile(savePath, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0644)
+	if err != nil {
+		Error("Download Error", "Error creating download file: "+err.Error())
+		return
+	}
+	defer f.Close()
+
+	done := false
+
+	go func(isDone *bool) {
+		// Update the bar display
+		for !*isDone {
+			dlModal.SetText(bar.String())
+			App.Draw()
+			time.Sleep(100 * time.Millisecond)
+		}
+	}(&done)
+
+	// Display
+	dlModal.ClearButtons()
+	dlModal.AddButtons([]string{"Downloading..."})
+	tabPages.ShowPage("dl")
+	tabPages.SendToFront("dl")
+	App.SetFocus(dlModal)
+	App.Draw()
+
+	io.Copy(io.MultiWriter(f, bar), resp.Body)
+	done = true
+	dlModal.SetText(fmt.Sprintf("Download complete! File saved to %s.", savePath))
+	dlModal.ClearButtons()
+	dlModal.AddButtons([]string{"Ok"})
+	dlModal.GetForm().SetFocus(100)
+	App.SetFocus(dlModal)
+	App.Draw()
+}
+
+// downloadPage saves the passed Page to a file.
+// It returns the saved path and an error.
+// It always cleans up, so if an error is returned there is no file saved
+func downloadPage(p *structs.Page) (string, error) {
+	var savePath string
+	var err error
+
+	if p.Mediatype == structs.TextGemini {
+		savePath, err = downloadNameFromURL(p.Url, ".gmi")
+	} else {
+		savePath, err = downloadNameFromURL(p.Url, ".txt")
+	}
+	if err != nil {
+		return "", err
+	}
+	err = ioutil.WriteFile(savePath, []byte(p.Raw), 0644)
+	if err != nil {
+		// Just in case
+		os.Remove(savePath)
+		return "", err
+	}
+	return savePath, err
+}
+
+// downloadNameFromURL takes a URl and returns a safe download path that will not overwrite any existing file.
+// ext is an extension that will be added if the file has no extension, and for domain only URLs.
+// It should include the dot.
+func downloadNameFromURL(u string, ext string) (string, error) {
+	var name string
+	var err error
+	parsed, _ := url.Parse(u)
+	if parsed.Path == "" || path.Base(parsed.Path) == "/" {
+		// No file, just the root domain
+		name, err = getSafeDownloadName(parsed.Hostname()+ext, true, 0)
+		if err != nil {
+			return "", err
+		}
+	} else {
+		// There's a specific file
+		name = path.Base(parsed.Path)
+		if !strings.Contains(name, ".") {
+			// No extension
+			name += ext
+		}
+		name, err = getSafeDownloadName(name, false, 0)
+		if err != nil {
+			return "", err
+		}
+	}
+	return filepath.Join(config.DownloadsDir, name), nil
+}
 
 // getSafeDownloadName is used by downloads.go only.
 // It returns a modified name that is unique for the downloads folder.
@@ -58,46 +266,4 @@ func getSafeDownloadName(name string, lastDot bool, n int) (string, error) {
 	}
 	d.Close()
 	return nn, nil // Name doesn't exist already
-}
-
-// downloadPage saves the passed Page to a file.
-// It returns the saved path and an error.
-// It always cleans up, so if an error is returned there is no file saved
-func downloadPage(p *structs.Page) (string, error) {
-	// Figure out file name
-	var name string
-	var err error
-	parsed, _ := url.Parse(p.Url)
-	if parsed.Path == "" || path.Base(parsed.Path) == "/" {
-		// No file, just the root domain
-		if p.Mediatype == structs.TextGemini {
-			name, err = getSafeDownloadName(parsed.Hostname()+".gmi", true, 0)
-			if err != nil {
-				return "", err
-			}
-		} else {
-			name, err = getSafeDownloadName(parsed.Hostname()+".txt", true, 0)
-			if err != nil {
-				return "", err
-			}
-		}
-	} else {
-		// There's a specific file
-		name = path.Base(parsed.Path)
-		if p.Mediatype == structs.TextGemini && !strings.HasSuffix(name, ".gmi") && !strings.HasSuffix(name, ".gemini") {
-			name += ".gmi"
-		}
-		name, err = getSafeDownloadName(name, false, 0)
-		if err != nil {
-			return "", err
-		}
-	}
-	savePath := filepath.Join(config.DownloadsDir, name)
-	err = ioutil.WriteFile(savePath, []byte(p.Raw), 0644)
-	if err != nil {
-		// Just in case
-		os.Remove(savePath)
-		return "", err
-	}
-	return savePath, err
 }